Back-End em andamento

+++ server/src/routes/memories.ts
@@ -1,58 +1,58 @@
-import { FastifyInstance } from "fastify";
-import { z } from "zod";
+import { FastifyInstance } from 'fastify'
+import { z } from 'zod'
 
-import { prisma } from "../lib/prisma";
+import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.addHook("preHandler", async (request) => {
-    await request.jwtVerify();
-  });
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
 
-  app.get("/memories", async (request) => {
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
       where: {
         userId: request.user.sub,
       },
       orderBy: {
-        createdAT: "asc",
+        createdAT: 'asc',
       },
-    });
+    })
 
     return memories.map((memory) => {
       return {
         id: memory.id,
         coverURL: memory.converURL,
-        excerpt: memory.content.substring(0, 120).concat("..."),
-      };
-    });
-  });
+        excerpt: memory.content.substring(0, 120).concat('...'),
+      }
+    })
+  })
 
-  app.get("/memories/:id", async (request, reply) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
-    });
+    })
 
     if (!memory.isPublic && memory.userId !== request.user.sub) {
-      return reply.status(401).send();
+      return reply.status(401).send()
     }
-  });
+  })
 
-  app.post("/memories", async (request) => {
+  app.post('/memories', async (request, reply) => {
     const bodySchema = z.object({
       content: z.string(),
       converURL: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    });
+    })
 
-    const { content, converURL, isPublic } = bodySchema.parse(request.body);
+    const { content, converURL, isPublic } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
       data: {
@@ -61,34 +61,34 @@ export async function memoriesRoutes(app: FastifyInstance) {
         isPublic,
         userId: request.user.sub,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.put("/memories/:id", async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const bodySchema = z.object({
       content: z.string(),
       converURL: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    });
+    })
 
-    const { content, converURL, isPublic } = bodySchema.parse(request.body);
+    const { content, converURL, isPublic } = bodySchema.parse(request.body)
 
     let memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
-    });
+    })
 
     if (memory.userId !== request.user.sub) {
-      return reply.status(401).send();
+      return reply.status(401).send()
     }
 
     memory = await prisma.memory.update({
@@ -100,22 +100,32 @@ export async function memoriesRoutes(app: FastifyInstance) {
         converURL,
         isPublic,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
         id,
       },
-    });
-  });
+    })
+  })
 }
